Reorganize hasher and add hasher benchmark
