+++ src/lib.rs
@@ -2,5 +2,7 @@
 #![feature(stdsimd)]
 
 mod gxhash;
+mod hasher;
 
 pub use gxhash::*;
+pub use hasher::*;
